% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmfmts.R
\name{cmfmts}
\alias{cmfmts}
\title{Calculates the complexity measures of a dataframe}
\usage{
cmfmts(
  dataset = NULL,
  n_cores = (parallel::detectCores() - 1),
  na = T,
  scale = T,
  timeout = 999999
)
}
\arguments{
\item{dataset}{Data to process (data.frame). Each row contains a time series.}

\item{n_cores}{Number of cores for parallel procesing (integer).}

\item{na}{Presence of NAs in the output dataframe (logical). It is recommended to use True in this parameter.}

\item{scale}{Standardization}

\item{timeout}{Maximum calculation time allowed per method.}
}
\value{
A dataframe containing the 41 time series characteristics for each time series processed.
}
\description{
Gets the complexity measures of a dataframe composed by
time series. Each row contains the values of a time series.
}
