% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{measure.aproximation_entropy}
\alias{measure.aproximation_entropy}
\title{Aproximation Entropy}
\usage{
measure.aproximation_entropy(time_serie, edim = 2, r = 0.2 * sd(time_serie))
}
\arguments{
\item{time_serie}{A time serie, object of type ts.}

\item{edim}{the embedding dimension, as for chaotic time series; a preferred value is 2}

\item{r}{filter factor; work on heart rate variability has suggested setting r to be 0.2 times the standard deviation of the data}
}
\value{
The value of the complexity measure: aproximation entropy.
}
\description{
Using package pracma (http://www.inside-r.org/packages/cran/pracma/docs/approx_entropy)
}
