% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nforbiden.R
\name{measure.PE_nforbiden}
\alias{measure.PE_nforbiden}
\title{forbiden patterns}
\usage{
measure.PE_nforbiden(time_serie, ndemb)
}
\arguments{
\item{time_serie}{input time series data}

\item{ndemb}{(OPTIONAL) If x is given, the embedding dimension (ndemb) is required.}
}
\value{
Number of "forbiden patterns" (cf. Amigo 2010)
}
\description{
This function calculates Number of "forbiden patterns" (cf. Amigo 2010)
Using package statcomp (https://www.rdocumentation.org/packages/statcomp/versions/0.0.1.1000/topics/global_complexity)
}
