% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{measure.permutation_entropy}
\alias{measure.permutation_entropy}
\title{Permutation Entropy}
\usage{
measure.permutation_entropy(time_serie, order = 4)
}
\arguments{
\item{time_serie}{A time serie, object of type ts.}

\item{order}{the order of the permutation entropy}
}
\value{
The value of the complexity measure: Permutation Entropy.
}
\description{
Using package tsExpKit (https://github.com/cbergmeir/tsExpKit/blob/master/R/dataComplexity.R)
}
