% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{measure.spectral_entropy}
\alias{measure.spectral_entropy}
\title{Spectral Entropy}
\usage{
measure.spectral_entropy(y)
}
\arguments{
\item{y}{input time series data}
}
\value{
spectral entropy
}
\description{
Calculation of the value of the spectral entropy.
Based on: http://uic.edu.hk/~kentsang/fyp2014/Fast%20Fourier%20Transform11.htm
}
